Rebase.

+++ sys/arch/powerpc/stand/mkbootimage/bebox_bootrec.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: bebox_bootrec.h,v 1.3 2012/12/24 14:01:24 kiyohara Exp $	*/
+/*	$NetBSD: bebox_bootrec.h,v 1.3.10.1 2014/08/10 06:54:06 tls Exp $	*/
 
 #define	BEBOX_HEADER_SIZE		0x6400
 #define	BEBOX_BLOCK_SIZE		0x200
@@ -136,9 +136,9 @@ u_char bebox_image_data12[] = {
 };
 
 struct bebox_image_block {
-	long offset;
+	int32_t offset;
 	u_char *data;
-	int size;
+	int32_t size;
 };
 
 #define	BEBOX_IMG(x)		__CONCAT(bebox_image_data,x)
@@ -162,7 +162,7 @@ struct bebox_image_block bebox_image_block[] = {
 	{ -1 }
 };
 
-long bebox_mtime_offset[] = {
+int32_t bebox_mtime_offset[] = {
 	0x00000004,
 	0x0000048c,
 	0x00000490,
